Move polyfill to a separate file, add some comments
